Small update
Reviews for the files now work. They are being posted and then are visible on the profile.
Profile reviews are implemented but profile is displayed as /app/account so I didn't know where to get the userid from however everything is implemented and is waiting for the userid to be in params.
Price duting upload is automatically capped at 99999 and minimum at 10.
Made api for pricechanging at /api/files/{fileid}/pricechange it posts price value and auth with token.